move executables into bin directory
