+++ .idea/runConfigurations/Run_Broker__8690__8691_.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Run Broker (8690, 8691)" type="NodeJSConfigurationType" application-parameters="configs/development-3.env.json" path-to-js-file="app.js" working-dir="$PROJECT_DIR$">
+  <configuration default="false" name="Run Broker (8690, 8691)" type="NodeJSConfigurationType" application-parameters="configs/development-3.env.json" path-to-js-file="bin/broker.js" working-dir="$PROJECT_DIR$">
     <envs>
       <env name="LOG_LEVEL" value="debug" />
     </envs>
