+++ Dockerfile
@@ -27,4 +27,4 @@ ENV LOG_LEVEL=info
 ENV CONFIG_FILE configs/docker-1.env.json
 ENV STREAMR_URL http://127.0.0.1:8081/streamr-core
 
-CMD node app.js ${CONFIG_FILE} --streamrUrl=${STREAMR_URL}
+CMD node bin/broker.js ${CONFIG_FILE} --streamrUrl=${STREAMR_URL}
